fixed rendering problem in App.tsx (fixed test) htbkoo/CssToAndFromReact
